Select, Delete, Insert, Table
SELECT FULLY OPERABLE.  Delete operable, but it requires the TransactionID rn, which would be too much to be displayed; I do not know if the select statement can call the ID without it being on the displayed table.  Insert statement was changed, but still inoperable.  The table was changed as well, as I found out the code I used from the internet was flawed.